Rename to fake IBAN; Add RequiredMembersTest
