+++ tests/_support/TestAsset/DummyHydratorAware.php
@@ -2,14 +2,14 @@
 
 declare(strict_types=1);
 
-namespace Codeception\Mock;
+namespace Codeception\TestAsset;
 
 use Zend\Hydrator\HydratorAwareInterface;
 use Zend\Hydrator\HydratorAwareTrait;
 
 /**
  * Class DummyHydratorAware
- * @package Codeception\Mock
+ * @package Codeception\TestAsset
  */
 
 class DummyHydratorAware implements HydratorAwareInterface
