+++ tests/unit/Hydrator/MandateTest.php
@@ -76,7 +76,7 @@ class MandateTest extends UnitTest
                 'bankAccount' => [
                     'owner' => 'PAY.',
                     'bic' => 'RABONL2U',
-                    'iban' => 'NL35RABO0117713678',
+                    'iban' => 'NL32RABO1670475085',
                 ],
             ],
             'amount' => [
@@ -143,7 +143,7 @@ class MandateTest extends UnitTest
                 'bankAccount' => [
                     'owner' => 'PAY.',
                     'bic' => 'RABONL2U',
-                    'iban' => 'NL35RABO0117713678',
+                    'iban' => 'NL32RABO1670475085',
                 ],
             ],
             'amount' => [
